Make a simple server API
Along with an annoying program to echo to clients.

The API is setup such that users of the class can decide whether
polling should be blocking or not.

I think this only works on Linux, haven't checked any other platforms.

There's "todo"s for error handling that I'll likely never actually do
unless I plan on using this for something "real".